requirements.txt
apscheduler==3.10.4
oauth2client==4.1.3
python-dotenv==1.0.0
pytz==2023.3
gspread==5.12.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
line-bot-sdk==3.6.0
fastapi==0.104.1
uvicorn==0.24.0
